+++ include/envoy/upstream/upstream.h
@@ -587,7 +587,9 @@ public:
   GAUGE(upstream_rq_pending_active, Accumulate)                                                    \
   GAUGE(version, NeverImport)                                                                      \
   HISTOGRAM(upstream_cx_connect_ms, Milliseconds)                                                  \
-  HISTOGRAM(upstream_cx_length_ms, Milliseconds)
+  HISTOGRAM(upstream_cx_length_ms, Milliseconds)                                                   \
+  HISTOGRAM(upstream_rq_timeout_budget_global, Unspecified)                                        \
+  HISTOGRAM(upstream_rq_timeout_budget_per_try, Unspecified)
 
 /**
  * All cluster load report stats. These are only use for EDS load reporting and not sent to the